last update before changing hero approach
