+++ app/components/HomeHero/FeaturedSongSlider/FeaturedSongSliderAnimationHelpers.ts
